pyproject.toml
[project]
name = "donna-life-os"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.19",
    "prompt-toolkit>=3.0.0",
    "pyyaml>=6.0",
    "rich>=14.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "watchdog>=6.0.0",
    "python-multipart>=0.0.18",
]
